+++ src/views/ModelPlan/CollaborationArea/Cards/ModelPlan/index.scss
